Refactor account controller and account services
