+++ src/main/java/uk/gov/hmcts/reform/pip/account/management/service/AccountService.java
@@ -1,36 +1,22 @@
 package uk.gov.hmcts.reform.pip.account.management.service;
 
-import com.microsoft.graph.models.User;
-import com.opencsv.bean.CsvToBean;
-import com.opencsv.bean.CsvToBeanBuilder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
+import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.pip.account.management.database.UserRepository;
 import uk.gov.hmcts.reform.pip.account.management.errorhandling.exceptions.AzureCustomException;
-import uk.gov.hmcts.reform.pip.account.management.errorhandling.exceptions.CsvParseException;
 import uk.gov.hmcts.reform.pip.account.management.errorhandling.exceptions.NotFoundException;
 import uk.gov.hmcts.reform.pip.account.management.errorhandling.exceptions.UserNotFoundException;
-import uk.gov.hmcts.reform.pip.account.management.helpers.EmailHelper;
-import uk.gov.hmcts.reform.pip.account.management.model.AzureAccount;
 import uk.gov.hmcts.reform.pip.account.management.model.CreationEnum;
 import uk.gov.hmcts.reform.pip.account.management.model.ListType;
-import uk.gov.hmcts.reform.pip.account.management.model.MediaCsv;
 import uk.gov.hmcts.reform.pip.account.management.model.PiUser;
 import uk.gov.hmcts.reform.pip.account.management.model.Roles;
 import uk.gov.hmcts.reform.pip.account.management.model.Sensitivity;
 import uk.gov.hmcts.reform.pip.account.management.model.UserProvenances;
-import uk.gov.hmcts.reform.pip.account.management.model.errored.ErroredAzureAccount;
 import uk.gov.hmcts.reform.pip.account.management.model.errored.ErroredPiUser;
 import uk.gov.hmcts.reform.pip.account.management.service.helpers.DateTimeHelper;
 import uk.gov.hmcts.reform.pip.model.enums.UserActions;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.io.Reader;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeParseException;
@@ -41,13 +27,10 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Stream;
 import javax.validation.ConstraintViolation;
 import javax.validation.Validator;
 
-import static uk.gov.hmcts.reform.pip.account.management.model.Roles.ALL_NON_RESTRICTED_ADMIN_ROLES;
-import static uk.gov.hmcts.reform.pip.account.management.model.Roles.ALL_NON_THIRD_PARTY_ROLES;
-import static uk.gov.hmcts.reform.pip.account.management.model.UserProvenances.ALL_NON_THIRD_PARTY_PROVENANCES;
+import static uk.gov.hmcts.reform.pip.account.management.model.UserProvenances.PI_AAD;
 import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
 
 /**
@@ -55,8 +38,7 @@ import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
  * The storage mechanism (e.g Azure) is seperated into a separate class.
  */
 @Slf4j
-@Component
-@SuppressWarnings({"PMD.LawOfDemeter", "PMD.ExcessiveImports", "PMD.TooManyMethods"})
+@Service
 public class AccountService {
 
     @Autowired
@@ -64,6 +46,9 @@ public class AccountService {
 
     @Autowired
     AzureUserService azureUserService;
+
+    @Autowired
+    AzureAccountService azureAccountService;
 
     @Autowired
     UserRepository userRepository;
@@ -79,85 +64,6 @@ public class AccountService {
 
     @Autowired
     SubscriptionService subscriptionService;
-
-    private static final String EMAIL_NOT_SENT_MESSAGE =
-        "Account has been successfully created, however email has failed to send.";
-
-    /**
-     * Method to create new accounts in azure.
-     *
-     * @param azureAccounts The accounts to be created.
-     * @param issuerId      The id of the user who created the accounts.
-     * @return Returns a map which contains two lists, Errored and Created accounts. Created will have object ID set.
-     **/
-    public Map<CreationEnum, List<? extends AzureAccount>> addAzureAccounts(//NOSONAR
-        List<AzureAccount> azureAccounts, String issuerId, boolean isExisting) {
-
-        Map<CreationEnum, List<? extends AzureAccount>> processedAccounts = new ConcurrentHashMap<>();
-
-        List<AzureAccount> createdAzureAccounts = new ArrayList<>();
-        List<ErroredAzureAccount> erroredAccounts = new ArrayList<>();
-
-        for (AzureAccount azureAccount : azureAccounts) {
-
-            Set<ConstraintViolation<AzureAccount>> constraintViolationSet = validator.validate(azureAccount);
-            if (!constraintViolationSet.isEmpty()) {
-
-                ErroredAzureAccount erroredSubscriber = new ErroredAzureAccount(azureAccount);
-                erroredSubscriber.setErrorMessages(constraintViolationSet
-                                                       .stream().map(constraint -> constraint.getPropertyPath()
-                        + ": " + constraint.getMessage()).toList());
-                erroredAccounts.add(erroredSubscriber);
-                continue;
-            }
-
-            try {
-
-                User userAzure = azureUserService.getUser(azureAccount.getEmail());
-
-                if (userAzure != null && !userAzure.givenName.isEmpty()
-                    && azureAccount.getRole().equals(Roles.VERIFIED)) {
-
-                    boolean emailSent = publicationService.sendNotificationEmailForDuplicateMediaAccount(
-                            azureAccount.getEmail(), userAzure.givenName);
-
-                    checkAndAddToErrorAccount(emailSent, azureAccount,
-                                              "Unable to send duplicate media account email",
-                                              erroredAccounts);
-                } else {
-
-                    User user = azureUserService.createUser(azureAccount);
-                    azureAccount.setAzureAccountId(user.id);
-                    createdAzureAccounts.add(azureAccount);
-
-                    log.info(writeLog(issuerId, UserActions.CREATE_ACCOUNT, azureAccount.getAzureAccountId()));
-                    boolean emailSent = handleAccountCreationEmail(azureAccount, user.givenName, isExisting);
-                    checkAndAddToErrorAccount(emailSent, azureAccount, EMAIL_NOT_SENT_MESSAGE, erroredAccounts);
-                }
-
-            } catch (AzureCustomException azureCustomException) {
-                log.error(writeLog(issuerId, UserActions.CREATE_ACCOUNT, azureAccount.getAzureAccountId()));
-                ErroredAzureAccount erroredSubscriber = new ErroredAzureAccount(azureAccount);
-                erroredSubscriber.setErrorMessages(List.of(azureCustomException.getMessage()));
-                erroredAccounts.add(erroredSubscriber);
-            }
-
-        }
-
-        processedAccounts.put(CreationEnum.CREATED_ACCOUNTS, createdAzureAccounts);
-        processedAccounts.put(CreationEnum.ERRORED_ACCOUNTS, erroredAccounts);
-
-        return processedAccounts;
-    }
-
-    private void checkAndAddToErrorAccount(boolean checkCondition, AzureAccount azureAccount, String errorMessage,
-                                           List<ErroredAzureAccount> erroredAccounts) {
-        if (!checkCondition) {
-            ErroredAzureAccount softErroredAccount = new ErroredAzureAccount(azureAccount);
-            softErroredAccount.setErrorMessages(List.of(errorMessage));
-            erroredAccounts.add(softErroredAccount);
-        }
-    }
 
     /**
      * Method to add users to P&I database, loops through the list and validates the email provided then adds them to
@@ -261,73 +167,6 @@ public class AccountService {
         return emailMap;
     }
 
-    public Map<CreationEnum, List<?>> uploadMediaFromCsv(MultipartFile mediaCsv, String issuerId) { //NOSONAR
-        List<MediaCsv> mediaList;
-
-        try (InputStreamReader inputStreamReader = new InputStreamReader(mediaCsv.getInputStream());
-             Reader reader = new BufferedReader(inputStreamReader)) {
-
-            CsvToBean<MediaCsv> csvToBean = new CsvToBeanBuilder<MediaCsv>(reader)
-                .withType(MediaCsv.class)
-                .build();
-
-            mediaList = csvToBean.parse();
-        } catch (Exception ex) {
-            throw new CsvParseException(ex.getMessage());
-        }
-
-        return addToAzureAndPiUsers(mediaList, issuerId);
-    }
-
-    private Map<CreationEnum, List<?>> addToAzureAndPiUsers(List<MediaCsv> accounts, String issuerId) {
-        Map<CreationEnum, List<? extends AzureAccount>> azureAccounts;
-        Map<CreationEnum, List<?>> piUserAccounts;
-        Map<CreationEnum, List<?>> completedAccounts = new ConcurrentHashMap<>();
-
-        azureAccounts = addAzureAccounts(accountModelMapperService.createAzureUsersFromCsv(accounts),
-                                         issuerId, true
-        );
-        piUserAccounts = addUsers(
-            accountModelMapperService
-                .createPiUsersFromAzureAccounts(azureAccounts.get(CreationEnum.CREATED_ACCOUNTS)),
-            issuerId
-        );
-
-
-        completedAccounts.put(
-            CreationEnum.CREATED_ACCOUNTS,
-            piUserAccounts.get(CreationEnum.CREATED_ACCOUNTS)
-        );
-        completedAccounts.put(
-            CreationEnum.ERRORED_ACCOUNTS,
-            Stream.concat(
-                azureAccounts.get(CreationEnum.ERRORED_ACCOUNTS).stream(),
-                piUserAccounts.get(CreationEnum.ERRORED_ACCOUNTS).stream()
-            ).distinct().toList()
-        );
-        return completedAccounts;
-    }
-
-    private boolean handleAccountCreationEmail(AzureAccount createdAccount, String fullName,
-                                               boolean isExisting) {
-        return switch (createdAccount.getRole()) {
-            case INTERNAL_ADMIN_CTSC, INTERNAL_ADMIN_LOCAL, INTERNAL_SUPER_ADMIN_CTSC, INTERNAL_SUPER_ADMIN_LOCAL ->
-                publicationService.sendNotificationEmail(createdAccount.getEmail(),
-                                                         createdAccount.getFirstName(), createdAccount.getSurname());
-            case VERIFIED ->  publicationService.sendMediaNotificationEmail(createdAccount.getEmail(),
-                                                                            fullName, isExisting);
-            default -> false;
-        };
-    }
-
-    public String getAccManDataForMiReporting() {
-        StringBuilder builder = new StringBuilder(54);
-        builder.append("user_id,provenance_user_id,user_provenance,roles").append(System.lineSeparator());
-        userRepository.getAccManDataForMI()
-            .forEach(line -> builder.append(line).append(System.lineSeparator()));
-        return builder.toString();
-    }
-
     /**
      * Delete a user account by the supplied email.
      * This deletes the user from AAD, our user table and subscriptions.
@@ -340,7 +179,7 @@ public class AccountService {
             .orElseThrow(() -> new NotFoundException("User with supplied ID could not be found"));
 
         try {
-            if (UserProvenances.PI_AAD.equals(userToDelete.getUserProvenance())) {
+            if (PI_AAD.equals(userToDelete.getUserProvenance())) {
                 azureUserService.deleteUser(userToDelete.getProvenanceUserId());
             }
         } catch (AzureCustomException ex) {
@@ -390,53 +229,6 @@ public class AccountService {
     }
 
     /**
-     * Request a page of users that can be filtered down with the supplied parameters.
-     *
-     * @param pageable The pageable object.
-     * @param email The email to query by.
-     * @param userProvenanceId The user provenance id to query by.
-     * @param userProvenances A list of user provenances to query by.
-     * @param roles A list of roles to query by.
-     * @param userId The user id to query by.
-     * @return A page with a list of piUsers.
-     */
-    public Page<PiUser> findAllAccountsExceptThirdParty(Pageable pageable, String email, String userProvenanceId,
-                                                        List<UserProvenances> userProvenances, List<Roles> roles,
-                                                        String userId) {
-
-        // If a user ID is supplied then only find by that
-        if (!userId.isBlank()) {
-            return userRepository.findByUserIdPageable(userId, pageable);
-        }
-
-        //  If a list of user provenances are supplied then use that else use all non third party user provenances
-        List<UserProvenances> userProvenancesToQuery = ALL_NON_THIRD_PARTY_PROVENANCES;
-        if (!userProvenances.isEmpty()) {
-            userProvenancesToQuery = userProvenances;
-        }
-
-        // If a list of roles are supplied then use that else use all non third party roles
-        List<Roles> rolesToQuery = ALL_NON_THIRD_PARTY_ROLES;
-        if (!roles.isEmpty()) {
-            rolesToQuery = roles;
-        }
-
-        // If user provenance id is supplied then find by an exact match
-        String userProvenanceIdToQuery = "%%";
-        if (!userProvenanceId.isBlank()) {
-            userProvenanceIdToQuery = userProvenanceId;
-        }
-
-        return userRepository.findAllByEmailLikeIgnoreCaseAndUserProvenanceInAndRolesInAndProvenanceUserIdLike(
-            "%" + email + "%",
-            userProvenancesToQuery,
-            rolesToQuery,
-            userProvenanceIdToQuery,
-            pageable
-        );
-    }
-
-    /**
      * Process updating a role for an account.
      *
      * @param userId The ID of the user to update.
@@ -448,7 +240,7 @@ public class AccountService {
             "User with supplied user id: %s could not be found", userId)));
 
         // If they are a PI AAD user then try update the users role in B2C
-        if (UserProvenances.PI_AAD.equals(userToUpdate.getUserProvenance())) {
+        if (PI_AAD.equals(userToUpdate.getUserProvenance())) {
             try {
                 azureUserService.updateUserRole(userToUpdate.getProvenanceUserId(), updatedRole.toString());
             } catch (AzureCustomException ex) {
@@ -466,15 +258,6 @@ public class AccountService {
     }
 
     /**
-     * Method which will retrieve all accounts which are third party.
-     *
-     * @return The list of found PiUser accounts.
-     */
-    public List<PiUser> findAllThirdPartyAccounts() {
-        return userRepository.findAllByUserProvenance(UserProvenances.THIRD_PARTY);
-    }
-
-    /**
      * This method retrieves a user by their ID.
      * @param userId The user ID to retrieve
      * @return The user that is found.
@@ -483,44 +266,5 @@ public class AccountService {
     public PiUser getUserById(UUID userId) {
         return userRepository.findByUserId(userId).orElseThrow(() -> new NotFoundException(String.format(
             "User with supplied user id: %s could not be found", userId)));
-    }
-
-    public AzureAccount retrieveAzureUser(String provenanceUserId) {
-        try {
-            Optional<PiUser> user = userRepository.findByProvenanceUserIdAndUserProvenance(
-                provenanceUserId, UserProvenances.PI_AAD);
-            if (user.isPresent()) {
-                AzureAccount azureUser = new AzureAccount();
-                User aadUser = azureUserService.getUser(user.get().getEmail());
-                azureUser.setAzureAccountId(aadUser.id);
-                azureUser.setFirstName(aadUser.givenName);
-                azureUser.setSurname(aadUser.surname);
-                azureUser.setDisplayName(aadUser.displayName);
-                azureUser.setEmail(user.get().getEmail());
-                return azureUser;
-            } else {
-                throw new NotFoundException(String.format(
-                    "User with supplied provenanceUserId: %s could not be found", provenanceUserId));
-            }
-        } catch (AzureCustomException e) {
-            log.error(writeLog(UUID.fromString(provenanceUserId), "Error while retrieving users details"));
-        }
-
-        throw new IllegalArgumentException("Error while retrieving user details with provenanceUserId: "
-                                               + provenanceUserId);
-    }
-
-    /**
-     * This method retrieves an admin user (excluding system admin) by their email and provenance.
-     * @param email The email of the user to retrieve
-     * @param provenance The provenance of the user to retrieve
-     * @return The user that is found.
-     * @throws NotFoundException if a user is not found.
-     */
-    public PiUser getAdminUserByEmailAndProvenance(String email, UserProvenances provenance) {
-        return userRepository.findByEmailIgnoreCaseAndUserProvenanceAndRolesIn(email, provenance,
-                                                                               ALL_NON_RESTRICTED_ADMIN_ROLES)
-            .orElseThrow(() -> new NotFoundException(String.format(
-                "No user found with the email: %s and provenance %s", EmailHelper.maskEmail(email), provenance)));
     }
 }
