+++ src/main/java/uk/gov/hmcts/reform/pip/account/management/validation/validator/ConditionalEmailValidator.java
@@ -9,12 +9,6 @@ import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
 public class ConditionalEmailValidator implements ConstraintValidator<PiEmailConditionalValidation, PiUser> {
-
-    @Override
-    public void initialize(PiEmailConditionalValidation email) {
-        // Do nothing
-    }
-
     /**
      * This is a note that the below validator wipes incoming email addresses when they are added erroneously to
      * third party creation requests. This prevents us having access to any data we shouldn't have. The suppression
@@ -25,7 +19,6 @@ public class ConditionalEmailValidator implements ConstraintValidator<PiEmailCon
      * @return boolean for whether it's successfully validated or not
      */
     @Override
-    @SuppressWarnings("PMD.LawOfDemeter")
     public boolean isValid(PiUser user, ConstraintValidatorContext context) {
         if (Roles.getAllThirdPartyRoles().contains(user.getRoles())) {
             user.setEmail(null);
