+++ src/main/java/uk/gov/hmcts/reform/pip/account/management/validation/validator/NotSystemAdminValidatorAzureAccount.java
@@ -1,20 +1,20 @@
 package uk.gov.hmcts.reform.pip.account.management.validation.validator;
 
 import uk.gov.hmcts.reform.pip.account.management.model.AzureAccount;
-import uk.gov.hmcts.reform.pip.account.management.model.Roles;
 import uk.gov.hmcts.reform.pip.account.management.validation.annotations.NotSystemAdmin;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
+import static uk.gov.hmcts.reform.pip.account.management.model.Roles.SYSTEM_ADMIN;
+
 /**
  * Validator class that checks if an azure account is a system admin.
  */
-@SuppressWarnings({"PMD.LawOfDemeter"})
 public class NotSystemAdminValidatorAzureAccount implements ConstraintValidator<NotSystemAdmin, AzureAccount> {
 
     @Override
     public boolean isValid(AzureAccount azureAccount, ConstraintValidatorContext constraintValidatorContext) {
-        return !Roles.SYSTEM_ADMIN.equals(azureAccount.getRole());
+        return !SYSTEM_ADMIN.equals(azureAccount.getRole());
     }
 }
