+++ src/test/java/uk/gov/hmcts/reform/pip/account/management/service/SystemAdminAccountServiceTest.java
@@ -44,7 +44,7 @@ class SystemAdminAccountServiceTest {
     private PublicationService publicationService;
 
     @Mock
-    private AccountService accountService;
+    private AzureAccountService azureAccountService;
 
     @Mock
     private UserRepository userRepository;
@@ -86,7 +86,7 @@ class SystemAdminAccountServiceTest {
 
         systemAdminAccountService = new SystemAdminAccountService(validator, azureUserService, userRepository,
                                                                   publicationService, 4,
-                                                                  accountService);
+                                                                  azureAccountService);
 
     }
 
@@ -98,7 +98,7 @@ class SystemAdminAccountServiceTest {
         when(azureUserService.createUser(argThat(user -> EMAIL.equals(user.getEmail()))))
             .thenReturn(expectedUser);
         when(userRepository.save(any())).thenReturn(expectedPiUser);
-        when(accountService.retrieveAzureUser(any()))
+        when(azureAccountService.retrieveAzureAccount(any()))
             .thenReturn(azUser);
         when(publicationService.sendNotificationEmail(EMAIL, FORENAME, SURNAME)).thenReturn(Boolean.TRUE);
         when(userRepository.findByUserId(any())).thenReturn(Optional.ofNullable(expectedPiUser));
@@ -117,7 +117,7 @@ class SystemAdminAccountServiceTest {
 
         when(userRepository.findByUserId(any()))
             .thenReturn(Optional.ofNullable(expectedPiUser));
-        when(accountService.retrieveAzureUser(any()))
+        when(azureAccountService.retrieveAzureAccount(any()))
             .thenReturn(azUser);
         when(azureUserService.createUser(argThat(user -> EMAIL.equals(user.getEmail()))))
             .thenThrow(new AzureCustomException("Test error"));
@@ -139,7 +139,7 @@ class SystemAdminAccountServiceTest {
 
         when(userRepository.findByUserId(any()))
             .thenReturn(Optional.ofNullable(expectedPiUser));
-        when(accountService.retrieveAzureUser(any()))
+        when(azureAccountService.retrieveAzureAccount(any()))
             .thenReturn(azUser);
         when(validator.validate(any(), any())).thenReturn(Set.of(constraintViolation));
         when(constraintViolation.getMessage()).thenReturn("This is a message");
@@ -199,7 +199,7 @@ class SystemAdminAccountServiceTest {
             .thenReturn(Optional.of(expectedPiUser));
         when(userRepository.findByUserId(any()))
             .thenReturn(Optional.ofNullable(expectedPiUser));
-        when(accountService.retrieveAzureUser(any()))
+        when(azureAccountService.retrieveAzureAccount(any()))
             .thenReturn(azUser);
 
         SystemAdminAccountException systemAdminAccountException =
@@ -218,7 +218,7 @@ class SystemAdminAccountServiceTest {
             .thenReturn(Optional.empty());
         when(userRepository.findByUserId(any()))
             .thenReturn(Optional.ofNullable(expectedPiUser));
-        when(accountService.retrieveAzureUser(any()))
+        when(azureAccountService.retrieveAzureAccount(any()))
             .thenReturn(azUser);
         when(userRepository.findByRoles(Roles.SYSTEM_ADMIN)).thenReturn(List.of(expectedPiUser, expectedPiUser,
                                                                                 expectedPiUser, expectedPiUser));
